requirements.txt
annotated-types==0.6.0
anyio==4.2.0
blinker==1.7.0
certifi==2023.11.17
click==8.1.7
deprecation==2.1.0
Flask==3.0.0
gotrue==2.1.0
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
packaging==23.2
postgrest==0.13.1
pydantic==2.5.3
pydantic_core==2.14.6
python-dateutil==2.8.2
python-dotenv==1.0.0
realtime==1.0.2
six==1.16.0
sniffio==1.3.0
storage3==0.7.0
StrEnum==0.4.15
supabase==2.3.1
supafunc==0.3.3
typing_extensions==4.9.0
websockets==11.0.3
Werkzeug==3.0.1
